requirements.txt
flask
youtube_search

pafy
python-vlc